Move some base components to subfolders
