+++ assets/js/base/components/cart-checkout/index.js
@@ -1,4 +1,10 @@
+export { default as AddressForm } from './address-form';
 export { default as Button } from './button';
+export { default as CheckoutForm } from './form';
+export { default as FormStep } from './form-step';
+export { default as NoShipping } from './no-shipping';
+export { default as PlaceOrderButton } from './place-order-button';
+export { default as Policies } from './policies';
 export { default as ProductImage } from './product-image';
 export { default as ProductLowStockBadge } from './product-low-stock-badge';
 export { default as ProductMetadata } from './product-metadata';
@@ -6,5 +12,8 @@ export { default as ProductName } from './product-name';
 export { default as ProductPrice } from './product-price';
 export { default as ProductSaleBadge } from './product-sale-badge';
 export { default as ProductVariationData } from './product-variation-data';
-export { default as PlaceOrderButton } from './place-order-button';
 export { default as ReturnToCartButton } from './return-to-cart-button';
+export { default as ShippingCalculator } from './shipping-calculator';
+export { default as ShippingLocation } from './shipping-location';
+export { default as ShippingRatesControl } from './shipping-rates-control';
+export * from './totals/index.js';
