+++ assets/js/base/components/cart-checkout/policies/index.js
