+++ assets/js/base/components/cart-checkout/policies/style.scss
