+++ assets/js/base/components/cart-checkout/shipping-calculator/address.js
@@ -3,7 +3,10 @@
  */
 import PropTypes from 'prop-types';
 import { __ } from '@wordpress/i18n';
-import { Button } from '@woocommerce/base-components/cart-checkout';
+import {
+	AddressForm,
+	Button,
+} from '@woocommerce/base-components/cart-checkout';
 import { useState } from '@wordpress/element';
 import isShallowEqual from '@wordpress/is-shallow-equal';
 import { useValidationContext } from '@woocommerce/base-context';
@@ -12,7 +15,6 @@ import { useValidationContext } from '@woocommerce/base-context';
  * Internal dependencies
  */
 import './style.scss';
-import AddressForm from '../address-form';
 
 const ShippingCalculatorAddress = ( {
 	address: initialAddress,
