+++ assets/js/base/components/cart-checkout/totals/totals-coupon-code-input/style.scss
