+++ assets/js/base/components/cart-checkout/totals/totals-footer-item/style.scss
