+++ assets/js/base/components/cart-checkout/totals/totals-shipping-item/has-shipping-rate.js
