--- assets/js/base/components/checkout/index.js
@@ -1,4 +1,0 @@
-export { default as CheckoutForm } from './form';
-export { default as FormStep } from './form-step';
-export { default as NoShipping } from './no-shipping';
-export { default as Policies } from './policies';
