+++ assets/js/base/components/reviews/index.js
@@ -1,0 +1,3 @@
+export { default as ReviewList } from './review-list';
+export { default as ReviewListItem } from './review-list-item';
+export { default as ReviewSortSelect } from './review-sort-select';
