+++ assets/js/blocks/reviews/editor-block.js
@@ -8,8 +8,10 @@ import { Disabled } from '@wordpress/components';
 import { REVIEW_RATINGS_ENABLED } from '@woocommerce/block-settings';
 import ErrorPlaceholder from '@woocommerce/block-components/error-placeholder';
 import LoadMoreButton from '@woocommerce/base-components/load-more-button';
-import ReviewList from '@woocommerce/base-components/review-list';
-import ReviewSortSelect from '@woocommerce/base-components/review-sort-select';
+import {
+	ReviewList,
+	ReviewSortSelect,
+} from '@woocommerce/base-components/reviews';
 import withReviews from '@woocommerce/base-hocs/with-reviews';
 
 /**
