+++ assets/js/blocks/reviews/frontend-block.js
@@ -6,8 +6,10 @@ import { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { REVIEW_RATINGS_ENABLED } from '@woocommerce/block-settings';
 import LoadMoreButton from '@woocommerce/base-components/load-more-button';
-import ReviewSortSelect from '@woocommerce/base-components/review-sort-select';
-import ReviewList from '@woocommerce/base-components/review-list';
+import {
+	ReviewList,
+	ReviewSortSelect,
+} from '@woocommerce/base-components/reviews';
 import withReviews from '@woocommerce/base-hocs/with-reviews';
 
 /**
